+++ Source/YaOpt/Patches/Verse_AI_AttackTargetFinder_BestAttackTarget.cs
@@ -11,7 +11,8 @@ namespace YaOpt.Patches
 	/// <seealso cref="YaOptSettings.OptAttackTargetFinder"/>
 	internal static class Verse_AI_AttackTargetFinder_BestAttackTarget
 	{
-		private static readonly Dictionary<int, bool> _validationCache = new Dictionary<int, bool>();
+		[ThreadStatic]
+		private static Dictionary<int, bool> _validationCache;
 
 		[HarmonyPatch(typeof(AttackTargetFinder))]
 		[HarmonyPatch(nameof(AttackTargetFinder.BestAttackTarget))]
@@ -25,6 +26,8 @@ namespace YaOpt.Patches
 
 			static void Prefix()
 			{
+				if (_validationCache == null)
+					_validationCache = new Dictionary<int, bool>();
 				_validationCache.Clear();
 			}
 
@@ -37,9 +40,6 @@ namespace YaOpt.Patches
 		[HarmonyPatch]
 		private static class ClosurePart
 		{
-			// JobPredictor.PredictDoConstantJob will use this concurrently.
-			private static GreedySpinLock _spinLock = new GreedySpinLock();
-
 			static MethodBase TargetMethod()
 			{
 				MethodInfo method = null;
@@ -75,30 +75,14 @@ namespace YaOpt.Patches
 
 			static bool Prefix(IAttackTarget __0, ref bool __result)
 			{
-				_spinLock.Enter();
-				try
-				{
-					return !_validationCache.TryGetValue(__0.Thing.thingIDNumber, out __result);
-				}
-				finally
-				{
-					_spinLock.Exit();
-				}
+				return !_validationCache.TryGetValue(__0.Thing.thingIDNumber, out __result);
 			}
 
 			static void Postfix(IAttackTarget __0, ref bool __result, bool __runOriginal)
 			{
 				if (__runOriginal)
 				{
-					_spinLock.Enter();
-					try
-					{
-						_validationCache[__0.Thing.thingIDNumber] = __result;
-					}
-					finally
-					{
-						_spinLock.Exit();
-					}
+					_validationCache[__0.Thing.thingIDNumber] = __result;
 				}
 			}
 		}