+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 * Fix: Exception during game launch when both YaOpt and Performance Fish are installed, but Performance Fish's TryFindBestIngredientsHelpers_InnerDelegate is disabled.
+* Fix: System.InvalidOperationException when both Target Finding Optimization and Constant Job Prediction are enabled.
 
 ## 1.1.0