+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 * Fix: Exception during game launch when both YaOpt and Performance Fish are installed, but Performance Fish's TryFindBestIngredientsHelpers_InnerDelegate is disabled.
 * Fix: System.InvalidOperationException when both Target Finding Optimization and Constant Job Prediction are enabled.
+* Fix: Thread-safe for ListerBuildingsRepairable.
 
 ## 1.1.0